refactor(tests) - Tests to fixtures - mergeSiblingVars (#757)
* refactor(tests) - Tests to fixtures - mergeSiblingVars

* rename folders and fix formatting issues

[skip publish]
