+++ packages/babel-plugin-transform-merge-sibling-variables/__tests__/fixtures/block-scoped-for/actual.js
@@ -1,0 +1,5 @@
+let i = 0;
+for (let x = 0; x < 10; x++) console.log(i + x);
+
+const j = 0;
+for (const x = 0; ; ) console.log(j + x);
