+++ packages/babel-plugin-transform-merge-sibling-variables/__tests__/fixtures/block-scoped-outside-for/actual.js
@@ -1,0 +1,3 @@
+let i = 0;
+let y = 0;
+for (let x = 0; x < 10; x++) console.log(i + x);
