+++ packages/babel-plugin-transform-merge-sibling-variables/__tests__/fixtures/block-scoped-outside-for/expected.js
@@ -1,0 +1,4 @@
+let i = 0,
+    y = 0;
+
+for (let x = 0; x < 10; x++) console.log(i + x);
