+++ packages/babel-plugin-transform-merge-sibling-variables/__tests__/fixtures/concat-var-for-loop/actual.js
@@ -1,0 +1,3 @@
+var i = 0;
+var j = 0;
+for (var x = 0; x < 10; x++) console.log(i + x);
