+++ packages/babel-plugin-transform-merge-sibling-variables/__tests__/fixtures/concat-var-for-loop/expected.js
@@ -1,0 +1,2 @@
+
+for (var i = 0, j = 0, x = 0; x < 10; x++) console.log(i + x);
