+++ packages/babel-plugin-transform-merge-sibling-variables/__tests__/fixtures/concat-var/actual.js
@@ -1,0 +1,3 @@
+var i = 0;
+var x = 0;
+var y = 0;
