+++ packages/babel-plugin-transform-merge-sibling-variables/__tests__/fixtures/concat-var/expected.js
@@ -1,0 +1,3 @@
+var i = 0,
+    x = 0,
+    y = 0;
