+++ packages/babel-plugin-transform-merge-sibling-variables/__tests__/fixtures/lift-destructuring-var/expected.js
@@ -1,0 +1,6 @@
+for (var i = 0; i < 0; i++) {
+  var [j] = jj();
+}
+for (var i = 0; i < 0; i++) {
+  var { j } = jj();
+}
