+++ packages/babel-plugin-transform-merge-sibling-variables/__tests__/fixtures/lift-different-declars/expected.js
@@ -1,0 +1,3 @@
+for (let i = 0; i < 0; i++) {
+  var i = 0;
+}
