+++ packages/babel-plugin-transform-merge-sibling-variables/__tests__/fixtures/lift-let-declars/expected.js
@@ -1,0 +1,3 @@
+for (let i = 0, j; i < 0; i++) {
+  j = jj();
+}
