+++ packages/babel-plugin-transform-merge-sibling-variables/__tests__/fixtures/lift-mult-declars/actual.js
@@ -1,0 +1,4 @@
+for (var i = 0; i < 0; i++) {
+  var i = 0,
+    k = 0;
+}
