+++ packages/babel-plugin-transform-merge-sibling-variables/__tests__/fixtures/lift-no-loop-intializer/actual.js
@@ -1,0 +1,2 @@
+for (;;) {}
+for (;;) var i = 0;
