+++ packages/babel-plugin-transform-merge-sibling-variables/__tests__/fixtures/lift-not-initialized/actual.js
@@ -1,0 +1,3 @@
+for (var i = 0; ; ) {
+  var i;
+}
