+++ packages/babel-plugin-transform-merge-sibling-variables/__tests__/fixtures/lift-var-declars/actual.js
@@ -1,0 +1,4 @@
+for (var i = 0; i < 0; i++) {
+  var j = jj();
+}
+for (var i = 0; ; ) var j = 0;
