+++ packages/babel-plugin-transform-merge-sibling-variables/__tests__/fixtures/lift-var-declars/expected.js
@@ -1,0 +1,4 @@
+for (var i = 0, j; i < 0; i++) {
+  j = jj();
+}
+for (var i = 0, j;;) j = 0;
